# docs/swagger.yaml
definitions:
...
  models.ProductVariantInput:
    properties:
      add_price:
        type: integer
      product_id:
        type: integer
      temperature:
        type: string
    type: object
...
  models.Response:
    properties:
      message:
        type: string
      result: {}
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /product-sizes/{product_id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Ambil semua size berdasarkan product ID
      tags:
      - product-size
  /product-variants:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product Variant Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ProductVariantInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Tambah variant baru
      tags:
      - product-variant
  /product-variants/{id}:
    delete:
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Hapus variant
      tags:
      - product-variant
    patch:
      consumes:
      - application/json
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product Variant Input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ProductVariantInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Edit variant
      tags:
      - product-variant
  /product-variants/{product_id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Ambil semua variant berdasarkan ID
      tags:
      - product-variant
  /product-variants/detail/{id}:
    get:
      parameters:
      - description: Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Ambil 1 variant berdasarkan ID
      tags:
      - product-variant
...
swagger: "2.0"