# internal/handlers/product_size.go
package handlers

import (
	"koda-b6-backend1/internal/models"
	"koda-b6-backend1/internal/service"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

type ProductSizeHandler struct {
	service *service.ProductSizeService
}

func NewProductSizeHandler(service *service.ProductSizeService) *ProductSizeHandler {
	return &ProductSizeHandler{
		service: service,
	}
}

// @Summary Ambil semua size berdasarkan product ID
// @Tags product-size
// @Produce json
// @Param product_id path int true "Product ID"
// @Success 200 {object} models.Response
// @Router /product-sizes/{product_id} [get]
func (h *ProductSizeHandler) GetByProductID(ctx *gin.Context) {
	productID, err := strconv.Atoi(ctx.Param("product_id"))
	if err != nil {
		ctx.JSON(http.StatusBadRequest, models.Response{
			Success: false,
			Message: "product id harus berupa angka",
		})
		return
	}

	sizes, err := h.service.GetByProductID(productID)
	if err != nil {
		ctx.JSON(http.StatusNotFound, models.Response{
			Success: false,
			Message: err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, models.Response{
		Success: true,
		Message: "success",
		Result:  sizes,
	})
}